Making fakeplate like blank Plate keys with exports executecommand /me and hotwire - why u will get keys when u hotwired car and stoped the engine?

+++ shared/utils.lua
@@ -78,9 +78,11 @@ function Utils.VehicleType(value)
         end
     end
 
-    local vehicleType = GetVehicleType(tempVehicle)
-
-    DeleteEntity(tempVehicle)
+    local vehicleType = 'automobile'
+    if DoesEntityExist(tempVehicle) then
+        vehicleType = GetVehicleType(tempVehicle)
+        DeleteEntity(tempVehicle)
+    end
 
     return vehicleType
 end
